requirements-dev.txt
# Development Requirements for ArrPy
# Install with: pip install -r requirements-dev.txt

# Core requirements
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
pytest-xdist>=3.0.0  # Parallel test execution

# Code quality
ruff>=0.1.0          # Fast Python linter
black>=23.0.0        # Code formatter
mypy>=1.0.0          # Static type checker
isort>=5.12.0        # Import sorting

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
sphinx-autodoc-typehints>=1.22.0
numpydoc>=1.5.0

# Profiling and benchmarking
memory-profiler>=0.60.0
line-profiler>=4.0.0
py-spy>=0.3.14       # Sampling profiler

# Build tools
wheel>=0.38.0
build>=0.10.0
twine>=4.0.0         # For PyPI uploads

# Optional C++ build requirements
pybind11>=2.10.0     # For C++ extensions
cmake>=3.18          # For building C++ code